requirements.txt
pyyaml
addict
tensorboard
tqdm
SimpleITK
matplotlib
batchgenerators
einops
timm
monai
